requirements.txt
psycopg2-binary # To connect to the transactional database, i.e PostgreSQL
python-dotenv   # To read .env file and inject environment variables.
openai          # Connect to OpenAI through the API and perform translation
pika            # Connect to RabbitMQ, works as task queue
redis           # Connect to Redis, works as cache
